requirements.txt
Flask==0.10.1
requests==2.4.1
RPi.GPIO==0.6.3
